requirements.txt
fastapi
uvicorn[standard]
xgboost-cpu
pandas
scikit-learn